adding Phoebe feature.

+++ Metro.Phoebe/Metro.Phoebe.Shares/Model/StockBarData.cs
@@ -13,9 +13,9 @@ namespace Metro.Phoebe.Shares.Model
         public double High { get; set; }
         public double Low { get; set; }
         public double Close { get; set; }
-        public long Volume { get; set; }
+        public double Volume { get; set; }
 
-        public StockBarData(DateTime date, double open, double high, double low, double close, long volumn)
+        public StockBarData(DateTime date, double open, double high, double low, double close, double volumn)
         {
             this.Date = date;
             this.Open = open;
